feat: #39 Delete CategoryThird by Root
관리자 -상품 카테고리 소분류 상품 존재하는지 제약 추가 후 제약에 걸리지 않으면 삭제 구현완료

+++ src/components/amdin/Category/DeleteThirdPopup.vue
@@ -36,6 +36,20 @@ const closeDeleteModal = () => {
 
 const deleteCategoryThird = async () => {
   try {
+    const productResponse = await fetch(`/api/admin/product/category/${props.currentThirdCode}`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+    const products = await productResponse.json();
+
+    if(products.length > 0) {
+      alert("카테고리를 삭제할 수 없습니다. 해당 카테고리를 사용하는 제품이 존재합니다.");
+      return;
+    }
+
+
     const response = await fetch(`/api/admin/category/third/delete/${props.currentThirdCode}?requesterAdminCode=1`, {
       method: 'DELETE',
       headers: {
@@ -57,6 +71,7 @@ const deleteCategoryThird = async () => {
 
 const closePopup = () => {
   emits('close');
+  window.location.reload();
 };
 
 </script>
